# Projeto_2_API_Biblioteca/src/model/dto/PessoaRequestsDto.ts
export class PessoaRequestsDto {
    id: number;
    name: string;
    email: string;

    constructor(id?: number, name?: string, email?: string) {
        this.id = id || 0;
        this.name = name || '';
        this.email = email || '';
    }
}